# [REVIEW] Fixes groupby performance issue