<!--

Thank you for contributing to cuDF :)

Here are some guidelines to help the review process go smoothly.

1. Please write a description in this text box of the changes that are being
   made.

2. Please ensure that you have written units tests for the changes made/features
   added.

3. There are CI checks in place to enforce that committed code follows our style
   and syntax standards. Please see our contribution guide in `CONTRIBUTING.MD`
   in the project root for more information about the checks we perform and how
   you can run them locally.

4. If you are closing an issue please use one of the automatic closing words as
   noted here: https://help.github.com/articles/closing-issues-using-keywords/

5. If your pull request is not ready for review but you want to make use of the
   continuous integration testing facilities please label it with `[WIP]`.

6. If your pull request is ready to be reviewed without requiring additional
   work on top of it, then remove the `[WIP]` label (if present) and replace
   it with `[REVIEW]`. If assistance is required to complete the functionality,
   for example when the C/C++ code of a feature is complete but Python bindings
   are still required, then add the label `[HELP-REQ]` so that others can triage
   and assist. The additional changes then can be implemented on top of the
   same PR. If the assistance is done by members of the rapidsAI team, then no
   additional actions are required by the creator of the original PR for this,
   otherwise the original author of the PR needs to give permission to the
   person(s) assisting to commit to their personal fork of the project. If that
   doesn't happen then a new PR based on the code of the original PR can be
   opened by the person assisting, which then will be the PR that will be
   merged.

7. Once all work has been done and review has taken place please do not add
   features or make changes out of the scope of those requested by the reviewer
   (doing this just add delays as already reviewed code ends up having to be
   re-reviewed/it is hard to tell what is new etc!). Further, please do not
   rebase your branch on master/force push/rewrite history, doing any of these
   causes the context of any comments made by reviewers to be lost. If
   conflicts occur against master they should be resolved by merging master
   into the branch used for making the pull request.

Many thanks in advance for your cooperation!

-->
Taking care of series which will have same indices as the dataframe and skipping the join.

Old version groupby: 
`%time res = df.groupby([str(i) for i in range(5)], sort = True)['5','6'].sum()`
```
CPU times: user 465 ms, sys: 132 ms, total: 597 ms
Wall time: 597 ms
```
`%time res = df.groupby([str(i) for i in range(5)], sort = False)['5','6'].sum()`
```
CPU times: user 62.2 ms, sys: 24.8 ms, total: 86.9 ms
Wall time: 85.4 ms
```
New Version groupby:
`%time res = df.groupby([str(i) for i in range(5)], sort = True)['5','6'].sum()`
```
CPU times: user 353 ms, sys: 96.1 ms, total: 449 ms
Wall time: 449 ms
```
`%time res = df.groupby([str(i) for i in range(5)], sort = False)['5','6'].sum()`
```
CPU times: user 68.8 ms, sys: 20.2 ms, total: 89 ms
Wall time: 87.7 ms
```
My system spec:
2 Quadro RTX 8000
48 GB RAM
Intel(R) Xeon(R) Gold 6128 CPU @ 3.40GHz

close #3057