@ayushdg 2019-10-11
Tested with the fix. Groupby performance is back to what is expected. Thanks @rgsl888prabhu! 😄